requirements.txt
fastapi
uvicorn
pymongo
pydantic
python-dotenv
pytest
httpx